Use more precise twig version detection
