+++ composer.json
@@ -10,17 +10,17 @@
     "pug/installer": "^0.1.0",
     "pug-php/pug": "^2.7.1 || ^3.0.0",
     "pug-php/pug-assets": "^1.0.1",
-    "symfony/framework-bundle": "^2.7 || ^3.0 || ^4.0",
-    "symfony/http-foundation": "^2.7 || ^3.0 || ^4.0",
-    "symfony/http-kernel": "^2.7 || ^3.0 || ^4.0",
-    "symfony/security-bundle": "^2.7 || ^3.0 || ^4.0",
-    "symfony/templating": "^2.7 || ^3.0 || ^4.0",
-    "symfony/twig-bridge": "^2.7 || ^3.0 || ^4.0"
+    "symfony/framework-bundle": "^2.7 || ^3.0 || ^4.0 || ^5.0",
+    "symfony/http-foundation": "^2.7 || ^3.0 || ^4.0 || ^5.0",
+    "symfony/http-kernel": "^2.7 || ^3.0 || ^4.0 || ^5.0",
+    "symfony/security-bundle": "^2.7 || ^3.0 || ^4.0 || ^5.0",
+    "symfony/templating": "^2.7 || ^3.0 || ^4.0 || ^5.0",
+    "symfony/twig-bridge": "^2.7 || ^3.0 || ^4.0 || ^5.0"
   },
   "require-dev": {
     "phpunit/phpunit": "^4.8.35 || ^5.7 || ~6.3.0",
     "symfony/phpunit-bridge": "^3.3.9, !=3.4.31",
-    "symfony/symfony": "^2.7 || ^3.0 || ^4.0",
+    "symfony/symfony": "^2.7 || ^3.0 || ^4.0 || ^5.0",
     "codeclimate/php-test-reporter": "dev-master",
     "composer/composer": ">=1.0.0"
   },
